A few fixes
Bug Fixes:
> 1. Food no longer adds a floating point health value (long decimals)
> 2. Mob drops for orc have been buffed
> 3. Block drops no longer have a chance to drop 0 exp/gold
To be added:
> 1. Goblins no longer drop golden carrots when damaging players in safe spawn

+++ core/src/main/java/net/siegerpg/siege/core/listeners/DeathListener.java
@@ -74,7 +74,7 @@ public class DeathListener implements Listener, Runnable {
             if (player.getWorld().equals(Core.plugin().getServer().getWorld("SiegeHub"))) return;
             Bukkit.getServer().getScheduler().scheduleSyncDelayedTask(Core.plugin(), () -> {
                 double bal = Math.round(VaultHook.econ.getBalance(player));
-                double newBal = Math.round(bal * 0.95);
+                int newBal = (int) Math.round(bal * 0.95);
                 VaultHook.econ.withdrawPlayer(player, bal);
                 VaultHook.econ.depositPlayer(player, newBal);
                 player.sendTitle(Utils.tacc("&c&lYou Died"), Utils.tacc("&7" + (bal - newBal) + " has been lost"), 1, 60, 1);
